requirements.txt
requests>=2.31.0
pyttsx3>=2.90
pygame>=2.5.2
faster-whisper>=0.9.0
chromadb==0.4.15
langchain>=0.1.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
langchain-community>=0.1.0
langchain-chroma>=0.0.1
hnswlib==0.7.0
